third_party/blink/web_tests/platform/generic/http/tests/filesystem/workers/resolve-url-expected.txt

[Worker] Tests using webkitResolveLocalFileSystemURL to obtain an Entry from a URL

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/resolve-uri.js
[Worker] * Resolving a generated URL.
PASS [Worker] expectedPath is actualPath
PASS [Worker] isFile is true
[Worker] * Resolving test file by hand
PASS [Worker] expectedPath is actualPath
PASS [Worker] isFile is true
[Worker] * Resolving a URL with the wrong security origin (domain)
PASS [Worker] "SecurityError" is errorName
[Worker] * Resolving a URL with the wrong security origin (port)
PASS [Worker] "SecurityError" is errorName
[Worker] * Resolving a URL with the wrong security origin (scheme)
PASS [Worker] "SecurityError" is errorName
[Worker] * Resolving a completely bogus URL.
PASS [Worker] "EncodingError" is errorName
[Worker] * Resolving a URL with the wrong protocol
PASS [Worker] "EncodingError" is errorName
[Worker] * Resolving a URL with no slash between type and file
PASS [Worker] "EncodingError" is errorName
[Worker] * Resolving a URL with no slash between protocol and type (bogus port)
PASS [Worker] "SecurityError" is errorName
[Worker] * Resolve a path using backslashes
PASS [Worker] expectedPath is actualPath
PASS [Worker] isFile is true
[Worker] * Resolve a directory
PASS [Worker] expectedPath is actualPath
PASS [Worker] isDirectory is true
[Worker] * Resolve a path using a trailing slash
PASS [Worker] expectedPath is actualPath
PASS [Worker] isDirectory is true
[Worker] * Resolving a persistent URL.
PASS [Worker] expectedPath is actualPath
PASS [Worker] isDirectory is true
[Worker] Finished running tests.
PASS successfullyParsed is true

TEST COMPLETE
